Revert "Sync dev/imageResizer with master (#1105)"
This reverts commit db7f15541f452f20296cce6c8b06a06e49c3671e.
