--- installer/MSIX/build_msix.ps1
@@ -1,1 +1,0 @@
-makeappx build /v /overwrite /f PackagingLayout.xml /id "PowerToys-x64" /op bin\
