+++ installer/MSIX/msix_reinstall.ps1
@@ -1,11 +1,6 @@
 taskkill /f /im explorer.exe
-
 Get-AppxPackage -Name 'PowerToys' | select -ExpandProperty "PackageFullName" | Remove-AppxPackage
-
-.\build_msix.ps1
+makeappx build /v /overwrite /f PackagingLayout.xml /id "PowerToys-x64" /op bin\
 signtool sign /debug /a /fd SHA256 /f PowerToys_TemporaryKey.pfx /p 12345 bin\PowerToys-x64.msix
-signtool sign /debug /a /fd SHA256 /f PowerToys_TemporaryKey.pfx /p 12345 bin\PowerToys.msixbundle
-
 Add-AppxPackage .\bin\PowerToys-x64.msix
-
 start $Env:windir\explorer.exe
