+++ src/common/common.vcxproj
@@ -97,8 +97,6 @@
     <ClInclude Include="d2d_text.h" />
     <ClInclude Include="d2d_window.h" />
     <ClInclude Include="dpi_aware.h" />
-    <ClInclude Include="window_helpers.h" />
-    <ClInclude Include="icon_helpers.h" />
     <ClInclude Include="hwnd_data_cache.h" />
     <ClInclude Include="json.h" />
     <ClInclude Include="monitors.h" />
@@ -132,12 +130,10 @@
     </ClCompile>
     <ClCompile Include="settings_helpers.cpp" />
     <ClCompile Include="settings_objects.cpp" />
-    <ClCompile Include="icon_helpers.cpp" />
     <ClCompile Include="start_visible.cpp" />
     <ClCompile Include="tasklist_positions.cpp" />
     <ClCompile Include="common.cpp" />
     <ClCompile Include="windows_colors.cpp" />
-    <ClCompile Include="window_helpers.cpp" />
     <ClCompile Include="winstore.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
