+++ src/common/common.vcxproj.filters
@@ -81,12 +81,6 @@
     <ClInclude Include="winstore.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="icon_helpers.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="window_helpers.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="d2d_svg.cpp">
@@ -136,12 +130,6 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="winstore.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="icon_helpers.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="window_helpers.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
