--- src/common/window_helpers.h
@@ -1,4 +1,0 @@
-#pragma once
-#include "common.h"
-
-HWND CreateMsgWindow(_In_ HINSTANCE hInst, _In_ WNDPROC pfnWndProc, _In_ void* p);
