+++ src/modules/fancyzones/lib/FancyZones.h
@@ -36,7 +36,6 @@ interface __declspec(uuid("{5C8D99D6-34B2-4F4A-A8E5-7483F6869775}")) IZoneWindow
     IFACEMETHOD_(void, MoveWindowsOnActiveZoneSetChange)() = 0;
     IFACEMETHOD_(COLORREF, GetZoneHighlightColor)() = 0;
     IFACEMETHOD_(GUID, GetCurrentMonitorZoneSetId)(HMONITOR monitor) = 0;
-    IFACEMETHOD_(int, GetZoneHighlightOpacity)() = 0;
 };
 
 winrt::com_ptr<IFancyZones> MakeFancyZones(HINSTANCE hinstance, IFancyZonesSettings* settings) noexcept;
