+++ src/modules/fancyzones/lib/Settings.h
@@ -15,7 +15,6 @@ struct Settings
     bool appLastZone_moveWindows = false;
     bool use_cursorpos_editor_startupscreen = true;
     std::wstring zoneHightlightColor = L"#0078D7";
-    int zoneHighlightOpacity = 90;
     PowerToysSettings::HotkeyObject editorHotkey = PowerToysSettings::HotkeyObject::from_settings(true, false, false, false, VK_OEM_3);
     std::wstring excludedApps = L"";
     std::vector<std::wstring> excludedAppsArray;
