+++ src/modules/fancyzones/lib/ZoneWindow.cpp
@@ -546,14 +546,14 @@ void ZoneWindow::DrawActiveZoneSet(wil::unique_hdc& hdc, RECT const& clientRect)
 
         //                                 { fillAlpha, fill, borderAlpha, border, thickness }
         ColorSetting const colorHints      { 225, RGB(81, 92, 107),   255, RGB(104, 118, 138), -2 };
-        ColorSetting       colorViewer     { OpacitySettingToAlpha(m_host->GetZoneHighlightOpacity()), 0, 255, RGB(40, 50, 60),    -2 };
-        ColorSetting       colorHighlight  { OpacitySettingToAlpha(m_host->GetZoneHighlightOpacity()), 0, 255, 0, -2 };
+        ColorSetting       colorViewer     { 225, 0,                  255, RGB(40, 50, 60),    -2 };
+        ColorSetting       colorHighlight  { 225, 0,                  255, 0,                  -2 };
         ColorSetting const colorFlash      { 200, RGB(81, 92, 107),   200, RGB(104, 118, 138), -2 };
 
         auto zones = m_activeZoneSet->GetZones();
         const size_t maxColorIndex = min(size(zones) - 1, size(colors) - 1);
         size_t colorIndex = maxColorIndex;
-        for (auto iter = zones.begin(); iter != zones.end(); iter++)
+        for (auto iter = zones.rbegin(); iter != zones.rend(); iter++)
         {
             winrt::com_ptr<IZone> zone = iter->try_as<IZone>();
             if (!zone)
