+++ src/modules/fancyzones/lib/resource.h
@@ -13,4 +13,3 @@
 #define IDS_SETTING_LAUNCH_EDITOR_DESCRIPTION 113
 #define IDS_SETTING_LAUNCH_EDITOR_HOTKEY_LABEL 114
 #define IDS_SETTING_EXCLCUDED_APPS_DESCRIPTION 115
-#define IDS_SETTINGS_HIGHLIGHT_OPACITY 116
