+++ src/modules/fancyzones/lib/util.h
@@ -133,9 +133,3 @@ inline void ParseDeviceId(PCWSTR deviceId, PWSTR parsedId, size_t size)
         StringCchCopy(parsedId, size, L"FallbackDevice");
     }
 }
-
-inline int OpacitySettingToAlpha(int opacity)
-{
-    // convert percentage to a 0-255 alpha value
-    return opacity * 2.55;
-}
