+++ src/modules/powerrename/dll/PowerRenameExt.cpp
@@ -5,7 +5,6 @@
 #include <PowerRenameManager.h>
 #include <trace.h>
 #include <Helpers.h>
-#include <icon_helpers.h>
 #include <Settings.h>
 #include "resource.h"
