+++ src/modules/powerrename/lib/PowerRenameManager.cpp
@@ -4,7 +4,6 @@
 #include <algorithm>
 #include <shlobj.h>
 #include "helpers.h"
-#include "window_helpers.h"
 #include <filesystem>
 #include "trace.h"
