+++ src/modules/powerrename/testapp/PowerRenameTest.vcxproj
@@ -185,11 +185,6 @@
   <ItemGroup>
     <ResourceCompile Include="PowerRenameTest.rc" />
   </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\common\common.vcxproj">
-      <Project>{74485049-c722-400f-abe5-86ac52d929b3}</Project>
-    </ProjectReference>
-  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
