+++ src/runner/restart_elevated.h
@@ -3,4 +3,3 @@ void schedule_restart_as_elevated();
 void schedule_restart_as_non_elevated();
 bool is_restart_scheduled();
 bool restart_if_scheduled();
-bool restart_same_elevation();
