+++ src/settings-web/package.json
@@ -29,8 +29,8 @@
   },
   "dependencies": {
     "@svgr/webpack": "^4.3.3",
-    "@uifabric/azure-themes": "^7.0.10",
-    "office-ui-fabric-react": "^7.76.2",
+    "@uifabric/azure-themes": "^7.0.9",
+    "office-ui-fabric-react": "^7.51.0",
     "react": "~16.8.0",
     "react-dom": "~16.8.0"
   },
